Android Mobile Application with Java
The aim is to develop a mobile application using several structures. In this project, Navigation Drawer, Recycler View and Image Slider were used together; it was developed in Java language in Android Studio environment.
The aim is to develop a mobile application using several structures. In this project, Navigation Drawer, Recycler View and Image Slider were used together; it was developed in Java language in Android Studio environment.